+++ app/class-integration.php
@@ -23,12 +23,20 @@ if ( ! defined( 'ABSPATH' ) ) {
 abstract class Integration {
 
 	/**
+	 * Track slug name.
+	 *
+	 * @var string $slug
+	 * @since 1.16.0
+	 */
+	protected $slug;
+
+	/**
 	 * Current working directory.
 	 *
 	 * @var string $dir
 	 * @since 1.0.0
 	 */
-	protected $dir = null;
+	protected $dir;
 
 	/**
 	 * Current working path.
@@ -38,7 +46,7 @@ abstract class Integration {
 	 * @var string $dir
 	 * @since 1.0.0
 	 */
-	protected $local_path = null;
+	protected $local_path;
 
 	/**
 	 * List of required dependencies.
@@ -46,7 +54,15 @@ abstract class Integration {
 	 * @var array $active
 	 * @since 1.0.0
 	 */
-	protected $dependencies = [];
+	protected $dependencies;
+
+	/**
+	 * Additional inline CSS configuration items.
+	 *
+	 * @var array $inline_css_configs
+	 * @since 1.16.0
+	 */
+	protected $inline_css_configs = [];
 
 	/**
 	 * Setup integration.
@@ -57,6 +73,7 @@ abstract class Integration {
 	 * @param array  $dependencies List of required dependencies.
 	 */
 	public function __construct( $slug, $dependencies ) {
+		$this->slug         = $slug;
 		$this->dependencies = $dependencies;
 		$this->local_path   = trailingslashit( '/app/integrations' ) . $slug;
 		$this->dir          = get_template_directory() . $this->get_local_path();
@@ -102,6 +119,23 @@ abstract class Integration {
 	 */
 	public function is_active() {
 		return ! in_array( false, $this->dependencies, true );
+	}
+
+	/**
+	 * Load inline CSS if any output controls are defined.
+	 *
+	 * @since 1.16.0
+	 */
+	public function inline_css_configs( $configs ) {
+		if ( empty( $this->inline_css_configs ) ) {
+			return $configs;
+		}
+
+		foreach ( $this->inline_css_configs as $key ) {
+			$configs[ $this->slug . '-' . $key ] = include $this->get_dir() . '/customize/output/' . $key . '.php';
+		}
+
+		return $configs;
 	}
 
 }