+++ app/integrations/facetwp/customize.php
@@ -11,6 +11,3 @@
 
 // Load panels/sections/controls.
 require_once bigbox_get_integration( 'facetwp' )->get_dir() . '/customize/controls/navbar.php';
-
-// Output.
-require_once bigbox_get_integration( 'facetwp' )->get_dir() . '/customize/output/refresh.php';