+++ app/integrations/facetwp/template-functions.php
@@ -27,7 +27,20 @@ function bigbox_facetwp_wp_enqueue_scripts() {
 		'jquery',
 	];
 
-	wp_enqueue_script( "{$stylesheet}-facetwp", get_template_directory_uri() . '/public/js/facetwp.min.js', $deps, $version, true );
+	wp_enqueue_script(
+		"{$stylesheet}-facetwp",
+		get_template_directory_uri() . '/public/js/facetwp.min.js',
+		$deps,
+		$version,
+		true
+	);
+
+	wp_enqueue_style(
+		"{$stylesheet}-facetwp",
+		get_template_directory_uri() . '/public/css/facetwp.min.css',
+		[],
+		$version
+	);
 }
 
 /**
@@ -50,7 +63,7 @@ function bigbox_facetwp_loop() {
  * @return string
  */
 function bigbox_facetwp_navbar_search() {
-	return bigbox_get_view( 'navbar-search', [], bigbox_get_integration( 'facetwp' )->get_local_path() . '/views/partials' );
+	return bigbox_get_view( 'integrations/facetwp/partials/navbar-search' );
 }
 
 /**