+++ app/integrations/woocommerce-product-vendors/template-hooks.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * WooCommerce Brands template hooks.
+ * WooCommerce Product Vendors template hooks.
  *
  * @since 1.14.0
  *
@@ -15,19 +15,11 @@ if ( ! defined( 'ABSPATH' ) ) {
 
 $wcpv = new WC_Product_Vendors_Vendor_Frontend();
 
-// Filter shortcode-registration-form.php template location.
-add_filter(
-	'woocommerce_locate_template',
-	function( $template, $template_name ) {
-		if ( 'shortcode-registration-form.php' === $template_name ) {
-			return get_theme_file_path( 'app/integrations/woocommerce-product-vendors/templates/shortcode-registration-form.php' );
-		};
+// Filter template loading.
+add_filter( 'woocommerce_locate_template', 'bigbox_woocommerce_product_vendors_locate_template', 10, 2 );
 
-		return $template;
-	},
-	10,
-	2
-);
+// Assets.
+add_action( 'wp_enqueue_scripts', 'bigbox_woocommerce_product_vendors_enqueue_styles' );
 
 // Change "Sold by" on loop.
 remove_class_action( 'woocommerce_after_shop_loop_item', 'WC_Product_Vendors_Vendor_Frontend', 'add_sold_by_loop', 9 );