+++ app/theme/assets.php
@@ -26,11 +26,21 @@ function bigbox_enqueue_styles() {
 	$google = bigbox_get_google_fonts_url();
 
 	if ( $google ) {
-		wp_enqueue_style( $stylesheet . '-fonts', $google, [], $version );
+		wp_enqueue_style(
+			$stylesheet . '-fonts',
+			$google,
+			[],
+			$version
+		);
 	}
 
 	// Base and dynamic styles.
-	wp_enqueue_style( $stylesheet, get_template_directory_uri() . '/public/css/app.min.css', [], $version );
+	wp_enqueue_style(
+		$stylesheet,
+		get_template_directory_uri() . '/public/css/app.min.css',
+		[],
+		$version
+	);
 	wp_style_add_data( $stylesheet, 'rtl', 'replace' );
 
 	/**
@@ -41,41 +51,13 @@ function bigbox_enqueue_styles() {
 	 * @param bool
 	 */
 	if ( apply_filters( 'bigbox_customize_css_inline', true ) ) {
-		wp_add_inline_style( $stylesheet, bigbox_customize_inline_css() );
+		wp_add_inline_style(
+			$stylesheet,
+			bigbox_customize_inline_css()
+		);
 	}
 }
-add_action( 'wp_enqueue_scripts', 'bigbox_enqueue_styles' );
-
-/**
- * Editor styles
- *
- * @since 1.0.0
- */
-function bigbox_editor_styles() {
-	$google = bigbox_get_google_fonts_url();
-	$styles = [ 'public/css/editor.min.css' ];
-
-	if ( $google ) {
-		$styles[] = $google;
-	}
-
-	add_editor_style( $styles );
-}
-add_action( 'after_setup_theme', 'bigbox_editor_styles' );
-
-// @codingStandardsIgnoreStart
-/**
- * Add type declarations to editor.
- *
- * @since 1.0.0
- */
-function bigbox_editor_inline_styles( $mceInit ) {
-	$mceInit['content_style'] = require_once get_template_directory() . '/app/customize/output/editor.php';
-
-	return $mceInit;
-}
-// @codingStandardsIgnoreEnd
-add_filter( 'tiny_mce_before_init', 'bigbox_editor_inline_styles' );
+add_action( 'wp_enqueue_scripts', 'bigbox_enqueue_styles', 20 );
 
 /**
  * Enqueue scripts.
@@ -91,11 +73,17 @@ function bigbox_enqueue_scripts() {
 	];
 
 	// Combined application scripts.
-	wp_enqueue_script( $stylesheet, get_template_directory_uri() . '/public/js/app.min.js', $deps, $version, true );
+	wp_enqueue_script(
+		$stylesheet,
+		get_template_directory_uri() . '/public/js/app.min.js',
+		$deps,
+		$version,
+		true
+	);
 
 	// Send information to application scripts.
 	$js_data = [
-		'backgroundColor' => sanitize_hex_color( get_background_color() ),
+		'backgroundColor' => sanitize_hex_color( maybe_hash_hex_color( get_background_color() ) ),
 	];
 
 	wp_localize_script(
@@ -111,4 +99,21 @@ function bigbox_enqueue_scripts() {
 		apply_filters( 'bigbox_js', $js_data )
 	);
 }
-add_action( 'wp_enqueue_scripts', 'bigbox_enqueue_scripts' );
+add_action( 'wp_enqueue_scripts', 'bigbox_enqueue_scripts', 20 );
+
+/**
+ * Fix skip link focus in IE11.
+ *
+ * This does not enqueue the script because it is tiny and because it is only for IE11,
+ * thus it does not warrant having an entire dedicated blocking script being loaded.
+ *
+ * @link https://git.io/vWdr2
+ *
+ * @since 1.16.0
+ */
+function bigbox_skip_link_focus_fix() {
+	echo '<script>';
+	echo file_get_contents( get_template_directory() . '/public/js/skip-link-focus-fix.min.js' );
+	echo '</script>';
+}
+add_action( 'wp_print_footer_scripts', 'bigbox_skip_link_focus_fix' );